2025-05-23
It's back, baby!

So I added back Date because one of our requirements is to USE association. Thing is though is that I saw that we have never touched the Settings class (the one with the only source of association) like.. ever. It's just floating there and we never used it. So, I decided to take the liberty of adding back Date (aka user's date of birth) and make it an association to Account. This is basically copying that one juicemind assignment we did but shush. I feel this is better and easier to implement than trying to use Settings and it's various association classes like Privacy and Security with our limited time on the assignment.

Hopefully this is okay for you! I have an idea for this so hang tight!